requirements.txt
gradio==6.9.0
GitPython==3.1.43
requests==2.31.0
ollama==0.3.1
openai==1.51.0
